Move commandline-header parsing code to utility module
as the code can be re-used in the update() implementation.
