+++ imag-store/src/create.rs
@@ -1,24 +1,20 @@
-use std::collections::BTreeMap;
 use std::path::PathBuf;
 use std::io::stdin;
 use std::fs::OpenOptions;
 use std::result::Result as RResult;
 use std::io::Read;
 use std::ops::DerefMut;
-use std::str::Split;
 
 use clap::ArgMatches;
-use toml::Table;
-use toml::Value;
 
 use libimagrt::runtime::Runtime;
 use libimagstore::store::Entry;
 use libimagstore::store::EntryHeader;
-use libimagutil::key_value_split::IntoKeyValue;
 
 use error::StoreError;
 use error::StoreErrorKind;
 use util::build_entry_path;
+use util::build_toml_header;
 
 type Result<T> = RResult<T, StoreError>;
 
@@ -126,61 +122,5 @@ fn entry_from_raw(raw_src: &str) -> String {
             .and_then(|mut f| f.read_to_string(&mut content));
     }
     content
-}
-
-fn build_toml_header(matches: &ArgMatches, header: EntryHeader) -> EntryHeader {
-    if let Some(headerspecs) = matches.values_of("header") {
-        let mut main = BTreeMap::new();
-        for tpl in headerspecs.into_iter().filter_map(|hs| String::from(hs).into_kv()) {
-            let (key, value) = tpl.into();
-            let mut split = key.split(".");
-            let current = split.next();
-            if current.is_some() {
-                insert_key_into(String::from(current.unwrap()), &mut split, value, &mut main);
-            }
-        }
-    }
-    header
-}
-
-fn insert_key_into(current: String,
-                   rest_path: &mut Split<&str>,
-                   value: String,
-                   map: &mut BTreeMap<String, Value>) {
-    let next = rest_path.next();
-
-    if next.is_none() {
-        map.insert(current, parse_value(value));
-    } else {
-        if map.contains_key(&current) {
-            match map.get_mut(&current).unwrap() {
-                &mut Value::Table(ref mut t) => {
-                    insert_key_into(String::from(next.unwrap()), rest_path, value, t);
-                },
-                _ => unreachable!(),
-            }
-        } else {
-            let mut submap = BTreeMap::new();
-            insert_key_into(String::from(next.unwrap()), rest_path, value, &mut submap);
-            map.insert(current, Value::Table(submap));
-        }
-    }
-}
-
-fn parse_value(value: String) -> Value {
-    fn is_ary(v: &String) -> bool {
-        v.chars().next() == Some('[') && v.chars().last() == Some(']') && v.len() >= 3
-    }
-
-    if value == "true" {
-        Value::Boolean(true)
-    } else if value == "false" {
-        Value::Boolean(false)
-    } else if is_ary(&value) {
-        let sub = &value[1..(value.len()-1)];
-        Value::Array(sub.split(",").map(|v| parse_value(String::from(v))).collect())
-    } else {
-        Value::String(value)
-    }
 }
